Pass empty string to hdparm --security-unlock in GenericHardwareManager
hdparm interprets the value, NULL, as a password with string value: "NULL".
This is in contrast to --security-set-pass, which interprets NULL as the
empty string. The hdparm man page details flags where NULL should be used,
but --security-unlock is not one of them.

Example output of hdparm with NULL password:

    [root@localhost ~]# hdparm --user-master u --security-unlock NULL /dev/sda
    security_password="NULL"

    /dev/sda:
    Issuing SECURITY_UNLOCK command, password="NULL", user=user
    SECURITY_UNLOCK: Input/output error

Example output of hdparm with "" as password:

    [root@localhost ~]# hdparm --user-master u --security-unlock "" /dev/sda
    security_password=""

    /dev/sda:
     Issuing SECURITY_UNLOCK command, password="", user=user

Note the value of security_password in both cases.

Change-Id: Ic24b706a04ff6c08d750b9e3d79eb79eab2952ad
Story: 2001762
Task: 12161