Do not abort ata secure erase if security is in enabled state
At least in the case of INTEL SSDSC2BX016T4R drives, a 'hdparm
--security-unlock' command does not remove the enabled line of the
security section of `hdparm -I`. As a consequence, the
GenericHardwareManager does not perform an ata secure erase (it relies
on this line to be set to `not enabled`).

For an INTEL SSDSC2BX016T4R, it has been manually verfied that the
following sequence of steps perform a secure erase sucessfully:

- hdparm --user-master u --security-unlock /dev/sda
- hdparm --user-master u --security-set-pass NULL /dev/sda
- hdparm --user-master u --security-erase-enhanced NULL /dev/sda

this is indicated by the "not enabled" line appearing in the output of
`hdparm -I /dev/sda`. It is therefore proposed that we do not actually
care if the `enabled` line appears in the output pre-erase. We should
only verify that is set to "not enabled" after an erase is attempted.

When we do not check the enabled line, the test
test_erase_block_device_ata_security_enabled becomes redundant with
test_erase_block_device_ata_security_enabled_unlock_exception taking its
place.

As the tests are very dependent on the order of calls to `utils.execute`, I
added some asserts to the failure cases to ensure that the correct call was
failing.

Change-Id: I5d4e18e7d7209dda89181fbcc202b199f130af89
Story: 2001763
Task: 12162